Implemented base Meta class
