+++ src/Models/PostMeta.php
@@ -12,18 +12,13 @@ use UnderScorer\ORM\Eloquent\Model;
  * @property string meta_key
  * @property mixed  meta_value
  */
-class PostMeta extends Model implements MetaInterface
+class PostMeta extends Meta
 {
 
     /**
      * @var bool
      */
     public $timestamps = false;
-
-    /**
-     * @var string
-     */
-    protected $primaryKey = 'meta_id';
 
     /**
      * @var array
@@ -41,65 +36,4 @@ class PostMeta extends Model implements MetaInterface
     {
         return $this->getConnection()->db->postmeta;
     }
-
-    /**
-     * @return mixed
-     */
-    public function getMetaValue()
-    {
-        return $this->meta_value;
-    }
-
-    /**
-     * @param mixed $value
-     *
-     * @return static
-     */
-    public function setMetaValue( $value )
-    {
-        $this->meta_value = $value;
-
-        return $this;
-    }
-
-    /**
-     * @param string $key
-     *
-     * @return static
-     */
-    public function setMetaKey( string $key )
-    {
-        $this->meta_key = $key;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getMetaKey(): string
-    {
-        return $this->meta_key;
-    }
-
-    /**
-     * @param $value
-     *
-     * @return void
-     */
-    public function setMetaValueAttribute( $value )
-    {
-        $this->attributes[ 'meta_value' ] = maybe_serialize( $value );
-    }
-
-    /**
-     * @param $value
-     *
-     * @return mixed
-     */
-    public function getMetaValueAttribute( $value )
-    {
-        return maybe_unserialize( $value );
-    }
-
 }
